requirements.txt
-e git://github.com/django/django.git@61fdb8d487f62e5b89b3bef9cb2b212dcec6d1e5#egg=Django-master
django-tables2==0.15.0
pocket==0.3.3
requests==2.2.1
six==1.6.1
wsgiref==0.1.2
